# swagger-new-endpoints.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/orders": {
      "post": {
        "operationId": "OrderController_create",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": { "example": "en" }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [true, false]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/CreateOrderDTO" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Orders"],
        "security": [{ "ACCESS Token": [] }],
        "summary": "Places a new order. Starts at status PENDING awaiting the store's accept/reject decision; driver-assignment search starts in parallel, not after the store decides.",
        "description": "Places a new order. Starts at status PENDING awaiting the store's accept/reject decision; driver-assignment search starts in parallel, not after the store decides."
      }
    },
    "/api/orders/{id}/unassign": {
      "patch": {
        "operationId": "OrderController_unassign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": { "example": "en" }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [true, false]
            }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Orders"],
        "security": [{ "ACCESS Token": [] }],
        "summary": "Removes the currently assigned driver from an order and resets its status back to READY_PICKUP so it can be re-offered.",
        "description": "Removes the currently assigned driver from an order and resets its status back to READY_PICKUP so it can be re-offered."
      }
    },
    "/api/orders/{id}/{status}": {
      "patch": {
        "operationId": "OrderController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          },
          {
            "name": "status",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "PENDING",
                "PREPARING",
                "READY_PICKUP",
                "ON_THE_WAY",
                "DELIVERED",
                "CANCELLED",
                "REJECTED",
                "PAYMENT_FAILD",
                "PENDING_PAYMENT"
              ],
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": { "example": "en" }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [true, false]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOrderStatusBodyDTO"
              }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Orders"],
        "security": [{ "ACCESS Token": [] }],
        "summary": "Generic order status transition (store: PREPARING / READY_PICKUP / REJECTED / CANCELLED; driver: ON_THE_WAY / DELIVERED). Driver transitions require lat/lng in the body and are geofence-checked against the store or customer address.",
        "description": "Generic order status transition (store: PREPARING / READY_PICKUP / REJECTED / CANCELLED; driver: ON_THE_WAY / DELIVERED). Driver transitions require lat/lng in the body and are geofence-checked against the store or customer address."
      }
    },
    "/api/orders/{id}/reject": {
      "patch": {
        "operationId": "OrderController_rejectAssignment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": { "example": "en" }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [true, false]
            }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Orders"],
        "security": [{ "ACCESS Token": [] }],
        "summary": "Driver declines an assignment offer. Benches the driver (AFK break, if enabled) and re-runs assignment for the order — next-nearest driver in AUTO mode, or left unassigned for the admin to pick in MANUAL mode.",
        "description": "Driver declines an assignment offer. Benches the driver (AFK break, if enabled) and re-runs assignment for the order — next-nearest driver in AUTO mode, or left unassigned for the admin to pick in MANUAL mode."
      }
    },
    "/api/orders/{id}/accept": {
      "patch": {
        "operationId": "OrderController_acceptAssignment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": { "example": "en" }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [true, false]
            }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Orders"],
        "security": [{ "ACCESS Token": [] }],
        "summary": "Driver accepts a pending assignment offer, binding them to the order. Fails if the store has already rejected/cancelled the order in the meantime.",
        "description": "Driver accepts a pending assignment offer, binding them to the order. Fails if the store has already rejected/cancelled the order in the meantime."
      }
    },
    "/api/orders/assign": {
      "patch": {
        "operationId": "OrderController_assign",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": { "example": "en" }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [true, false]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AssignOrderDTO" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Orders"],
        "security": [{ "ACCESS Token": [] }],
        "summary": "Admin manually assigns one or more orders to a specific driver — used in MANUAL assignment mode, or to override AUTO.",
        "description": "Admin manually assigns one or more orders to a specific driver — used in MANUAL assignment mode, or to override AUTO."
      }
    },
...
  },
  "info": {
    "title": "Makanak — New & Fixed Endpoints",
    "description": "Standalone subset of the full API, containing only the endpoints added or behaviorally changed during this engagement — for isolated testing.",
    "version": "1.0",
    "contact": {
      "name": "Mahmoud Elamrosy",
      "url": "www.linkedin.com/in/",
      "email": "melamrosy72@gmail.com"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "ACCESS Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "REFRESH Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "PASSWORD_RESET Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "VERIFY Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "VISITOR Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOrderDTO": {
        "type": "object",
        "properties": {
          "couponCode": { "type": "string" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OrderItemDTO" }
          },
          "bundleSelections": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BundleSelectionDTO" }
          },
          "addressId": { "type": "number" },
          "branchId": { "type": "number" },
          "tip": { "type": "number" },
          "type": {
            "type": "string",
            "enum": ["DELIVERY", "PICKUP", "CUSTOM_DELIVERY"]
          },
          "fortuneRewardId": { "type": "number" },
          "note": { "type": "string" },
          "paymentMethod": { "type": "string", "enum": ["CASH", "WALLET"] },
          "paidWithWallet": { "type": "boolean", "example": true },
          "isGift": { "type": "boolean", "example": true },
          "category": { "type": "string", "enum": ["IMMEDIATE", "SCHEDULED"] },
          "scheduledAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-07T09:28:46.658Z"
          },
          "transferNumber": { "type": "string" },
          "transferImage": { "type": "string", "format": "binary" }
        },
        "required": ["branchId", "paymentMethod"]
      },
      "ChangeOrderStatusBodyDTO": {
        "type": "object",
        "properties": {
          "lat": { "type": "number" },
          "lng": { "type": "number" }
        }
      },
      "AssignOrderDTO": {
        "type": "object",
        "properties": {
          "specialistId": { "type": "number" },
          "orderIds": {
            "example": [11111, 22222, 33333],
            "type": "array",
            "items": { "type": "number" }
          }
        },
        "required": ["specialistId", "orderIds"]
      },
...
      "OrderItemDTO": {
        "type": "object",
        "properties": {
          "serviceId": { "type": "number" },
          "sizeId": { "type": "number" },
          "addonIds": { "type": "array", "items": { "type": "string" } },
          "quantity": { "type": "number" }
        },
        "required": ["serviceId", "quantity"]
      },
      "BundleSelectionDTO": {
        "type": "object",
        "properties": {
          "bundleId": { "type": "number" },
          "paidItems": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BundleLineDTO" }
          },
          "freeItems": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BundleLineDTO" }
          }
        },
        "required": ["bundleId", "paidItems", "freeItems"]
      },
      "BundleLineDTO": {
        "type": "object",
        "properties": {
          "serviceId": { "type": "number" },
          "sizeId": { "type": "number" },
          "addonIds": { "type": "array", "items": { "type": "string" } },
          "quantity": { "type": "number" }
        },
        "required": ["serviceId"]
      }
    }
  }
}